+++ ios/Classes/PermissionManager.swift
@@ -22,6 +22,13 @@ class PermissionManager: NSObject {
         result(Codec.encodePermissionStatus(permissionStatus: permissionStatus))
     }
     
+    static func checkServiceStatus(permission: PermissionGroup, result: @escaping FlutterResult) {
+        let permissionStrategy = PermissionManager.createPermissionStrategy(permission: permission)
+        let serviceStatus = permissionStrategy.checkServiceStatus(permission: permission)
+        
+        result(Codec.encodeServiceStatus(serviceStatus: serviceStatus))
+    }
+    
     func requestPermissions(permissions: [PermissionGroup], completion: @escaping PermissionRequestCompletion) {
         var requestQueue = Set(permissions.map { $0 })
         var permissionStatusResult: [PermissionGroup: PermissionStatus] = [:]