+++ lib/permission_enums.dart
@@ -18,6 +18,21 @@ enum PermissionStatus {
   unknown
 }
 
+/// Defines the state of a service related to the permission group
+enum ServiceStatus {
+  /// The unknown service status indicates the state of the service could not be determined.
+  unknown,
+
+  /// There is no service for the supplied permission group.
+  notApplicable,
+
+  /// The service for the supplied permission group is disabled.
+  disabled,
+
+  /// The service for the supplied permission group is enabled.
+  enabled
+}
+
 /// Defines the permission groups for which permissions can be checked or requested.
 enum PermissionGroup {
   /// The unknown permission only used for return type, never requested