+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.2.0
+
+* Added new method `checkServiceStatus` to allow users to check if the location services (on Android and iOS) and motion services (iOS only) are enabled;
+* When checking permission status (using `checkPermissionStatus`) return `PermissionStatus.disabled` when permissions are granted or denied and the location services (on Android and iOS) or the motion services (iOS only) are disabled.
+
 ## 2.1.3
 
 * Fixed bug on iOS where result of the `openAppSettings` call always returned `false`;